+++ NEWS.md
@@ -1,3 +1,10 @@
+* some variable names updated
+* bug fixed in `summary_interact()`
+* add some tests
+* Added support for `tibble`s.
+* Added `tidy_summary()` generics function and method for **Hmisc**
+  `summary()` with `method = "reverse"`.
+
 # depigner 0.0.2
 
 * Fixed `summary_interact()` function.